Main higlight line now has a delay if multiple objects are selected, and slightly lighter lines are drawn from the objects to their average middle, before the main line is drawn from there to the mouse point.
Movement on command!!!  Added impulse to atoms that are selected when clicking in empty space.

Removed gravity effect from Plane to more easily demonstrate the effect of applying an impulse.

Found that applying force is accumulative to bodies, so they keep feeling that force.  That might work for environment effects, but any effect that burns energy should only apply impulse in exchange.