Large update - Energy now gets absorbed by Atoms and created into a Charge.
Besides the change stated above, there was a lot of cleaning up of the code.  Tried to make it more consistent, with better error checking, and debug messaging.

Charge class added to hold the charge that an Atom has when being aligned by a Player.

Also added Shell class that will eventually protect a Charge from dissipating into Space when exposed to the elements.

(This also made me think of the idea of a directional wind, so that only energy directly behind a shell will not evaporate)