I'm pretty sure I fixed the problem of too many physical objects being created in general.  Also haven't encountered an instance of the shapes flipping out after all the energy dissipates.  I think shapes were being created without bodies, or the other way around.
So to avoid the problem of collisions with nonexistent bodies, before a shape and body are removed, their game_object is set to None.  Then in the collision detection function, shapes with no game_object are removed from the Plane (again possibly).  If for some reason, they are not being removed from the space then there should be some kind of Plane.force_remove() created to handle this.