Started to work on Views interface to make better use of PyGame.Surface class.
View base class is built out quite a bit.  Considering making MainView handle opening up the PyGame.display.set_mode(): Surface instead of the Interface.  Interface will still be responsible for the PyGame.clock to keep track of time.  CameraView is in progress.  This will enable better scrolling than the garbage I have (presumably), as well as zooming.  GameView is going to be a big one and is currently jumbled.

I'm finally tired.