More work on creating Views and trying to progress with a system that I feel will be the most efficient.
All calls to a View.update() can include an optional Area rect which only updates the subviews in that region and blits that segment of the surface.  Even now if we have a huge surface for a game level, we should never need to write to all of it and then lose most of it when it gets cropped to the size of the screen.  Basically, I'm trying to avoid redrawing the entire map on every frame, only to crop out the tiny pieces that are used for the Main View.