Trying to focus on the GameView as that is the first piece I think I should test.
From there I will try to build up the CameraView so that I can scroll and zoom.  After that I can try plugging it all into the MainView.

GameView is going to contain the settings for how everything in the game should be seen.  So that means it is now going to handle Player/energy color, atom border color, atom border opacity, atom border width, etc.  Moving all of that stuff over should be quick.  I just have to step it up.