PhysicalWorld - So made the Plane be replaced by the PhysicalWorld.  Also, all of the calls to pymunk were moved to the PhysicalWorld library.
SpaceTime - This will be more responsible for the ability to go back and forth between events in the timeline.  I will make it responsible for tracking all the world events.  Essentially, I should always be able to paint an exact picture of what is going now by building up from the last checkpoint with the events list.  It will also be responsible for creating the atoms and energy in the world that the Game will load from the Level Data.

Player - Will now have unique hashes to compare and control energies.