GameView - More work to get GameView to be responsible for displaying everything.
Made it hold a dict of display_objects that will handle more fluid displays of data than the game world often can provide.  One example is its current use, the Shell, which slowly fades in and fades out after every 1/6 mark rather than always accurately representing the exact charge of the Shell.

Element - Will also generate a unique hash and store it.  This id is used in dicts like the display_objects in GameView.

Walls - Added displaying of Walls to GameView and changed the name of the prior Wall class to just Line.  Then added a Wall to the elements that will create lines, and that is about it for now.